Adding support for environment variables.
We want to add a layer of abstraction to pass username/password/token et al. The first idea was to use direct integration to keyvault. But we decided to de-couple this feature. The variables can be extracted outside of the tool and can be passed in as environment variables. To that extent, this gist will extend the language to support environment variables. In the variable declaration, you can push ~ and when this happens, the code will read it from environment variables (sys.env for Scala)